+++ android/app/src/main/java/com/on/turip/ui/common/mapper/VideoInfomationMapper.kt
@@ -1,7 +1,7 @@
 package com.on.turip.ui.common.mapper
 
 import com.on.turip.domain.content.video.VideoInformation
-import com.on.turip.ui.search.model.VideoInformationModel
+import com.on.turip.ui.compose.search.model.VideoInformationModel
 
 fun VideoInformation.toUiModel(): VideoInformationModel =
     VideoInformationModel(
