+++ android/app/src/main/java/com/on/turip/ui/compose/search/component/SearchResultList.kt
@@ -35,7 +35,7 @@ import com.on.turip.R
 import com.on.turip.ui.common.TuripUrlConverter.convertVideoThumbnailUrl
 import com.on.turip.ui.compose.designsystem.theme.TuripColor
 import com.on.turip.ui.compose.designsystem.theme.TuripTheme
-import com.on.turip.ui.search.model.VideoInformationModel
+import com.on.turip.ui.compose.search.model.VideoInformationModel
 import kotlinx.collections.immutable.ImmutableList
 
 @Composable
@@ -96,14 +96,17 @@ private fun SearchResultItem(
                 .padding(
                     horizontal = TuripTheme.spacing.extraLarge,
                     vertical = TuripTheme.spacing.small,
-                ).background(
+                )
+                .background(
                     color = TuripColor.LightGray01,
                     shape = itemShape,
-                ).border(
+                )
+                .border(
                     width = 1.dp,
                     color = TuripColor.Black06,
                     shape = itemShape,
-                ).clip(itemShape)
+                )
+                .clip(itemShape)
                 .clickable { onItemClick() }
                 .padding(TuripTheme.spacing.medium),
     ) {
