+++ android/app/src/main/java/com/on/turip/ui/compose/search/keyword/SearchScreen.kt
@@ -28,9 +28,6 @@ import com.on.turip.ui.compose.search.component.SearchResultList
 import com.on.turip.ui.compose.search.keyword.component.SearchAppBar
 import com.on.turip.ui.compose.search.keyword.component.SearchEmptyView
 import com.on.turip.ui.compose.search.keyword.component.SearchHistoryList
-import com.on.turip.ui.search.keywordresult.SearchUiEffect
-import com.on.turip.ui.search.keywordresult.SearchUiState
-import com.on.turip.ui.search.keywordresult.SearchViewModel
 import kotlinx.collections.immutable.persistentListOf
 
 @Composable
