+++ android/app/src/main/java/com/on/turip/ui/compose/search/regionresult/RegionResultUiEffect.kt
@@ -1,4 +1,4 @@
-package com.on.turip.ui.search.regionresult
+package com.on.turip.ui.compose.search.regionresult
 
 sealed interface RegionResultUiEffect {
     data object NavigateToLogin : RegionResultUiEffect
