+++ android/app/src/main/java/com/on/turip/ui/search/regionresult/RegionResultActivity.kt
@@ -9,6 +9,7 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import com.on.turip.ui.compose.designsystem.theme.TuripTheme
 import com.on.turip.ui.compose.search.regionresult.RegionResultScreen
+import com.on.turip.ui.compose.search.regionresult.RegionResultViewModel
 import com.on.turip.ui.login.LoginActivity
 import com.on.turip.ui.trip.TripDetailActivity
 import dagger.hilt.android.AndroidEntryPoint
